+++ example/android/app/src/main/kotlin/com/panther_nine/touch_ripple_effect_example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.panther_nine.touch_ripple_effect_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}