+++ lib/touch_ripple_effect.dart
@@ -1,2 +1,0 @@
-export './src/touch_feedback.dart';
-export './src/touch_ripple_effect.dart';